+++ lib/startback/event/agent.rb
@@ -1,0 +1,8 @@
+module Startback
+  class Event
+    class Agent
+      include Support::OperationRunner
+
+    end # class Agent
+  end # class Event
+end # module Starback